responses/pyproject.toml
[project]
name = "responses"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-projects>=2.0.0",
    "azure-identity>=1.26.0b2",
    "jupyter>=1.1.1",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
]
